requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.4
Django==4.1
django-cors-headers==4.5.0
django-debug-toolbar==4.3.0
django-webpush==0.3.5
djangorestframework==3.15.1
drf-yasg==1.21.10
http_ece==1.1.0
idna==3.7
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.2.2
py-vapid==1.9.0
pycparser==2.21
PyMySQL==1.1.0
pyTelegramBotAPI==4.14.0
pytz==2025.1
pywebpush==1.9.4
PyYAML==6.0.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
telebot==0.0.5
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
